componentize criteria box
